Fixing flying ball bug.
