+++ Pinball/Assets/Scripts/Behaviours/ResetBehaviour.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
-using UnityEditor;
 using UnityEngine;
 
 public class ResetBehaviour : MonoBehaviour
@@ -87,31 +86,3 @@ public class ResetBehaviour : MonoBehaviour
 
 }
 
-[CustomEditor(typeof(ResetBehaviour))]
-public class ResetBehaviourEditor : Editor
-{
-
-    public override void OnInspectorGUI()
-    {
-        ResetBehaviour myResetBehaviour = (ResetBehaviour)target;
-
-        myResetBehaviour.resetTrigger = EditorGUILayout.Toggle("Reset Trigger?", myResetBehaviour.resetTrigger);
-
-        if(myResetBehaviour.resetTrigger)
-        {
-            myResetBehaviour.objectWithTrigger = (GameObject)EditorGUILayout.ObjectField("Object with trigger",
-                myResetBehaviour.objectWithTrigger, typeof(GameObject), true);
-
-            myResetBehaviour.resetTriggerTo = EditorGUILayout.Toggle("Reset Trigger To:", myResetBehaviour.resetTriggerTo);
-        }
-
-        myResetBehaviour.resetPosition = EditorGUILayout.Toggle("Reset Position?", myResetBehaviour.resetPosition);
-
-        if(myResetBehaviour.resetPosition)
-        {
-            myResetBehaviour.objectToResetPosition =
-                (GameObject)EditorGUILayout.ObjectField("Object to be reseted", myResetBehaviour.objectToResetPosition, typeof(GameObject), true);
-        }
-
-    }
-}
