+++ Pinball/Assets/Scripts/Helpers/ResetBehaviourEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f50ec6260c962d442bbfb9b217fcd5d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
