+++ Pinball/ProjectSettings/EditorBuildSettings.asset
@@ -4,5 +4,8 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes: []
+  m_Scenes:
+  - enabled: 1
+    path: Assets/Scenes/FGArcadeScene.unity
+    guid: 33e24284d379c774ca00ce21d837244c
   m_configObjects: {}
